\name{internalfunction}
\alias{.basicGCplot}
\alias{.findNeutraLoss.scan}
\alias{.projectpdf}
\alias{.dial}
\alias{.dialog.mehalocoa.CDF2RData}
\alias{.dialog.mehalocoa.cldetect}
\title{Internal functions for the MeHaloCoA package.}
\description{
	Internal functions for the MeHaloCoA package
}
\details{
	These are not to be called by the user.
}
\author{
Yann GUITTON
}

